+++ frontend/src/components/panels/Insight/Graph/GraphView.css
