+++ frontend/src/components/panels/Insight/Graph/GraphViewForFullscreen.css
