+++ frontend/src/components/panels/Insight/Graph/GraphViewStandalone.jsx
