+++ frontend/src/components/panels/Insight/Graph/GraphViewWithModal.css
