+++ frontend/src/components/panels/Insight/Memo/MemoEditor.css
