+++ frontend/src/components/panels/Insight/Memo/MemoList.css
