+++ frontend/src/components/panels/Insight/Memo/MemoPanel.css
