+++ frontend/src/components/panels/Insight/hooks/useAudioRecorder.js
@@ -1,5 +1,5 @@
 import { useState, useRef } from 'react';
-import { transcribeAudio } from '../../../backend/api/backend';
+import { transcribeAudio } from '../../../../../../backend/api/backend';
 
 export default function useAudioRecorder(onTranscribe) {
