+++ frontend/src/components/panels/Project/NewBrainModal.jsx
@@ -1,6 +1,6 @@
 // src/components/NewBrainModal.jsx
 import React, { useEffect, useState, useRef } from 'react';
-import { createBrain } from '../../../backend/api/backend';
+import { createBrain } from '../../../../../backend/api/backend';
 import './NewBrainModal.css';
 import { RiDeleteBack2Line } from "react-icons/ri";
 export default function NewBrainModal({ onClose, onCreated }) {
