+++ frontend/src/components/panels/Source/FileView.jsx
@@ -2,17 +2,17 @@
 import React, { useState, useEffect } from 'react'
 import { pdfjs } from 'react-pdf';
 import workerSrc from 'pdfjs-dist/build/pdf.worker.min?url';
-import './styles/Common.css'
-import './styles/SourcePanel.css'
-import './styles/Scrollbar.css'
-import './styles/FileView.css'
+import '../../common/Common.css';
+import './SourcePanel.css';
+import '../styles/Scrollbar.css';
+import './FileView.css';
 import FileIcon from './FileIcon'
 import { TiUpload } from 'react-icons/ti'
 import { GoPencil } from 'react-icons/go';
 import { RiDeleteBinLine } from 'react-icons/ri';
-import { processText, deleteDB } from '../../api/graphApi';
-import { fetchGraphData } from '../../api/graphApi';
-import ConfirmDialog from '../common/ConfirmDialog'
+import { processText, deleteDB } from '../../../api/graphApi';
+import { fetchGraphData } from '../../../api/graphApi';
+import ConfirmDialog from '../../common/ConfirmDialog';
 import { AiOutlineLoading3Quarters } from 'react-icons/ai'
 import { AiOutlineNodeIndex } from "react-icons/ai";
 import {
@@ -34,7 +34,7 @@ import {
   createTextToGraph,
   uploadTextfiles,
   getNodesBySourceId
-} from '../../../../backend/api/backend'
+} from '../../../../../backend/api/backend'
 import { ToastContainer, toast } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
 import fileHandlers from './fileHandlers/fileHandlers';
