+++ frontend/src/components/panels/Source/HighlightPopup.jsx
