+++ frontend/src/components/panels/Source/PDFSidebar.css
