+++ frontend/src/components/panels/Source/SourceUploadModal.jsx
@@ -1,12 +1,13 @@
 // src/components/panels/SourceUploadModal.jsx
 import React, { useState, useRef } from 'react';
-import './styles/SourceUploadModal.css';
+import './SourceUploadModal.css';
 import { IoCloudUploadOutline } from "react-icons/io5";
 import { AiOutlineLoading3Quarters } from 'react-icons/ai';
 import FileIcon from './FileIcon';
 import {
   uploadPdfs, createTextFile, createTextToGraph
-} from '../../../../backend/api/backend';
+} from '../../../../../backend/api/backend';
+
 //import { pdfjs } from 'pdfjs-dist';
 //import workerSrc from 'pdfjs-dist/build/pdf.worker.min?url';
 import SourceQuotaBar from './SourceQuotaBar';
